refactor(Exception) : 정리
