+++ src/main/java/wooteco/subway/ExceptionController.java
@@ -7,13 +7,13 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.exception.DuplicateStationException;
-import wooteco.subway.exception.InvalidDistanceException;
-import wooteco.subway.exception.InvalidSectionOnLineException;
+import wooteco.subway.exception.section.DuplicateStationException;
+import wooteco.subway.exception.section.InvalidDistanceException;
+import wooteco.subway.exception.section.InvalidSectionOnLineException;
 import wooteco.subway.exception.NotFoundException;
-import wooteco.subway.exception.NotFoundLineException;
-import wooteco.subway.exception.NotFoundStationException;
-import wooteco.subway.exception.NullColorException;
+import wooteco.subway.exception.line.NotFoundLineException;
+import wooteco.subway.exception.station.NotFoundStationException;
+import wooteco.subway.exception.line.NullColorException;
 import wooteco.subway.exception.NullException;
 import wooteco.subway.exception.NullIdException;
 import wooteco.subway.exception.NullNameException;
