+++ src/main/java/wooteco/subway/exception/line/DuplicateLineException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.line;
 
 public class DuplicateLineException extends RuntimeException {
