+++ src/main/java/wooteco/subway/exception/line/NotFoundLineException.java
@@ -1,4 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.line;
+
+import wooteco.subway.exception.NotFoundException;
 
 public class NotFoundLineException extends NotFoundException {
