+++ src/main/java/wooteco/subway/exception/section/InvalidDistanceException.java
@@ -1,4 +1,6 @@
-package wooteco.subway.exception;
+package wooteco.subway.exception.section;
+
+import wooteco.subway.exception.SubwayException;
 
 public class InvalidDistanceException extends SubwayException {
