+++ src/main/java/wooteco/subway/section/SectionService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import javax.validation.Valid;
 import org.springframework.stereotype.Service;
-import wooteco.subway.exception.InvalidSectionOnLineException;
-import wooteco.subway.exception.NotFoundStationException;
+import wooteco.subway.exception.section.InvalidSectionOnLineException;
+import wooteco.subway.exception.station.NotFoundStationException;
 import wooteco.subway.section.dao.SectionDao;
 import wooteco.subway.section.dto.DeleteStationDto;
 import wooteco.subway.section.dto.SectionServiceDto;
