+++ src/main/java/wooteco/subway/station/StationControllerAdvice.java
@@ -4,7 +4,7 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.exception.NotFoundStationException;
+import wooteco.subway.exception.station.NotFoundStationException;
 
 @RestControllerAdvice
 public class StationControllerAdvice {
