+++ src/test/java/wooteco/subway/section/SectionServiceTest.java
@@ -10,8 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.jdbc.Sql;
-import wooteco.subway.exception.InvalidDistanceException;
-import wooteco.subway.exception.NotFoundStationException;
+import wooteco.subway.exception.section.InvalidDistanceException;
+import wooteco.subway.exception.station.NotFoundStationException;
 import wooteco.subway.line.dao.LineDao;
 import wooteco.subway.section.dao.SectionDao;
 import wooteco.subway.section.dto.DeleteStationDto;
